:recycle: refactor message type
